django_admin/app/requirements.txt
Django==4.2.11
psycopg2-binary==2.9.10
python-dotenv==1.0.1
split_settings==1.0.0
django-storages==1.14.4 
requests==2.32.3
boto3==1.34.2
django-debug-toolbar==4.4.6
split_settings==1.0.0
uwsgi==2.0.28
PyJWT==2.9.0
celery==5.4.0
flower==2.0.0
django-summernote==0.8.20.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
djangorestframework==3.15.2
whitenoise==5.2.0